+++ src/main/java/io/dataspaceconnector/service/resource/type/ArtifactService.java
@@ -429,6 +429,10 @@ public class ArtifactService extends BaseEntityService<Artifact, ArtifactDesc>
     }
 
     private InputStream toInputStream(final byte[] data) {
-        return new ByteArrayInputStream(data);
+        if (data != null) {
+            return new ByteArrayInputStream(data);
+        } else {
+            return ByteArrayInputStream.nullInputStream();
+        }
     }
 }