+++ src/test/java/io/dataspaceconnector/model/artifact/ArtifactFactoryTest.java
@@ -690,6 +690,24 @@ public class ArtifactFactoryTest {
     }
 
     @Test
+    public void updateByteSize_setByteFromNullToEmpty_willNotUpdate() {
+        /* ARRANGE */
+        final var artifact = factory.create(new ArtifactDesc());
+
+        /* ACT */
+        final byte[] data = {};
+        final var checksum = new CRC32C();
+        checksum.update(data, 0, data.length);
+        final var result = factory.updateByteSize(artifact, data);
+
+        /* ASSERT */
+        //bytesize and checksum should still be 0
+        assertFalse(result);
+        assertEquals(0, artifact.getByteSize());
+        assertEquals(checksum.getValue(), artifact.getCheckSum());
+    }
+
+    @Test
     public void updateByteSize_hasChanged_willUpdateByteSizeAndChecksum() {
         /* ARRANGE */
         final var artifact = factory.create(new ArtifactDesc());